Clarify wxAstroCapture bundles separately licenced libraries.
e.g.  is under the wxWidgets licence, an LGPL variant.

e.g.  (part of the Driver Library of wxAstroCapture)
is currently under the LGPL v3 or later, although Martin Burri and
Carsten Arnholm have said by email changing this to GPL v2 or later
to match the rest of wxAstroCapture is fine:
http://groups.yahoo.com/neo/groups/wxAstroCapture/conversations/messages/2098